+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ## [Unreleased]
 
+- Improve ergonomics of `VariableByteInteger`
+
 ## [0.8.0] - 2026-08-09
 
 - Allow inheritance from packet classes